attempted 2 of Q3

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,18 +8,19 @@
 int main(void)
 {
 	int i;
-	int first_fib = 0;
-	int sec_fib = 1;
+	long int first_fib = 0;
+	long int sec_fib = 1;
 
 	for (i = 0; i < 50; i++)
 	{
-		int new_fib = first_fib + sec_fib;
+		long int new_fib = first_fib + sec_fib;
+
 		first_fib = sec_fib;
 		sec_fib = new_fib;
 		if (i < 49)
-			printf("%i, ", new_fib);	
+			printf("%li, ", new_fib);
 		else
-			printf("%i", new_fib);
+			printf("%li", new_fib);
 	}
 	printf("\n");
 	return (0);
